% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metagenome.R
\name{add_genomespot}
\alias{add_genomespot}
\title{Add GenomeSPOT predictions}
\usage{
add_genomespot(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project}.}
}
\value{
Updated \code{gm_project} with a \code{genomespot} table list.
}
\description{
Add GenomeSPOT predictions
}
