% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metagenome.R
\name{add_nonpareil}
\alias{add_nonpareil}
\title{Add Nonpareil curves and summaries}
\usage{
add_nonpareil(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project}.}
}
\value{
Updated \code{gm_project} with a \code{nonpareil} table list.
}
\description{
Add Nonpareil curves and summaries
}
