% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metagenome.R
\name{add_sylph}
\alias{add_sylph}
\title{Add sylph profiles}
\usage{
add_sylph(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project}.}
}
\value{
Updated \code{gm_project}.
}
\description{
Adds \code{sylph_taxonomic} (taxonomic abundance) and \code{sylph_sequence} (sequence abundance)
genome-level profiles.
}
