% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{analysis_metadata}
\alias{analysis_metadata}
\title{Metadata for analyses, with excluded samples removed}
\usage{
analysis_metadata(project.l, include_excluded = FALSE)
}
\arguments{
\item{project.l}{A \code{gm_project}.}

\item{include_excluded}{Keep samples excluded from analyses (\code{Excluded}: config \code{exclude_column} or
\code{exclude_samples}); default \code{FALSE}.}
}
\value{
Data frame.
}
\description{
Metadata for analyses, with excluded samples removed
}
