% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{assign_shapes}
\alias{assign_shapes}
\title{Assign point shapes to the values of a variable}
\usage{
assign_shapes(values.v)
}
\arguments{
\item{values.v}{Vector (factor levels respected).}
}
\value{
Named integer vector of filled shapes (21 to 25, then open shapes).
}
\description{
Assign point shapes to the values of a variable
}
