% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{combined_profile_names}
\alias{combined_profile_names}
\title{Profiles available for study and comparison samples together}
\usage{
combined_profile_names(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project} with comparison samples.}
}
\value{
Character vector of names for \code{\link[=combined_profile]{combined_profile()}}.
}
\description{
Profiles available for study and comparison samples together
}
