% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2m}
\alias{df2m}
\title{Convert a data frame to a matrix using one column as row names}
\usage{
df2m(input.df, rownames_column = 1)
}
\arguments{
\item{input.df}{Data frame.}

\item{rownames_column}{Column (name or index) holding the row names.}
}
\value{
Numeric matrix.
}
\description{
Convert a data frame to a matrix using one column as row names
}
