% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{figure_path}
\alias{figure_path}
\title{Figure path in the project figure directory using the configured format}
\usage{
figure_path(config.l, ...)
}
\arguments{
\item{config.l}{A \code{gm_config}.}

\item{...}{Path components; the last is the file name without extension.}
}
\value{
Character path.
}
\description{
Figure path in the project figure directory using the configured format
}
