% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_registry.R
\name{has_output}
\alias{has_output}
\title{Test whether a pipeline output is available}
\usage{
has_output(config.l, key)
}
\arguments{
\item{config.l}{A \code{gm_config}.}

\item{key}{Registry key.}
}
\value{
Logical.
}
\description{
Test whether a pipeline output is available
}
