% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m2df}
\alias{m2df}
\title{Convert a matrix to a data frame, moving row names into the first column}
\usage{
m2df(input.m, column_name = "Feature")
}
\arguments{
\item{input.m}{Matrix.}

\item{column_name}{Name of the new first column.}
}
\value{
Data frame.
}
\description{
Convert a matrix to a data frame, moving row names into the first column
}
