% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_community.R
\name{permanova_caption}
\alias{permanova_caption}
\title{Short PERMANOVA and PERMDISP caption for figures}
\usage{
permanova_caption(permanova.df, permdisp.l = NULL)
}
\arguments{
\item{permanova.df}{Result of \code{\link[=run_permanova]{run_permanova()}}.}

\item{permdisp.l}{Optional result of \code{\link[=run_permdisp]{run_permdisp()}}.}
}
\value{
Character caption.
}
\description{
Short PERMANOVA and PERMDISP caption for figures
}
