% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot top loadings of a PCA axis}
\usage{
plot_loadings(ordination, axis = "PC1", n = 10)
}
\arguments{
\item{ordination}{A PCA \code{gm_ordination}.}

\item{axis}{Axis name.}

\item{n}{Number of features from each end.}
}
\value{
A ggplot.
}
\description{
Plot top loadings of a PCA axis
}
