% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_elements.R
\name{plot_mobile_prevalence}
\alias{plot_mobile_prevalence}
\title{Plot mobile element prevalence}
\usage{
plot_mobile_prevalence(prevalence.df, n_samples, title = NULL)
}
\arguments{
\item{prevalence.df}{\code{prevalence} from \code{\link[=mobile_element_summary]{mobile_element_summary()}}.}

\item{n_samples}{Number of samples analysed.}

\item{title}{Plot title.}
}
\value{
A ggplot.
}
\description{
Plot mobile element prevalence
}
