% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_plots.R
\name{plot_pangenome_categories}
\alias{plot_pangenome_categories}
\title{Plot pangenome gene category counts}
\usage{
plot_pangenome_categories(profile)
}
\arguments{
\item{profile}{Pangenome profile from \code{\link[=classify_pangenome]{classify_pangenome()}}.}
}
\value{
A ggplot.
}
\description{
Plot pangenome gene category counts
}
