% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{primary_group}
\alias{primary_group}
\title{First configured group variable}
\usage{
primary_group(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project}.}
}
\value{
Column name, or \code{NULL} when no group variable is configured.
}
\description{
First configured group variable
}
