% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_labels}
\alias{profile_labels}
\title{Display labels of profile features, named by feature ID}
\usage{
profile_labels(profile)
}
\arguments{
\item{profile}{A \code{gm_profile}.}
}
\value{
Named character vector.
}
\description{
Display labels of profile features, named by feature ID
}
