% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_samples}
\alias{profile_samples}
\title{Samples in a profile}
\usage{
profile_samples(profile)
}
\arguments{
\item{profile}{A \code{gm_profile}.}
}
\value{
Character vector.
}
\description{
Samples in a profile
}
