% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{project_path}
\alias{project_path}
\title{Resolve a path relative to the project directory}
\usage{
project_path(config.l, ...)
}
\arguments{
\item{config.l}{A \code{gm_config}.}

\item{...}{Path components; an absolute first component is used as is.}
}
\value{
Character path.
}
\description{
Resolve a path relative to the project directory
}
