% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isolate.R
\name{read_bakta_stats}
\alias{read_bakta_stats}
\title{Read Bakta annotation statistics}
\usage{
read_bakta_stats(path)
}
\arguments{
\item{path}{Path to \code{bakta_annotation_stats.tsv}.}
}
\value{
Data frame with \code{Genome} first.
}
\description{
Read Bakta annotation statistics
}
