% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mags.R
\name{read_checkm1}
\alias{read_checkm1}
\title{Read the CheckM1 extended QA table}
\usage{
read_checkm1(path)
}
\arguments{
\item{path}{Path to \code{checkm_qa_results.tsv}.}
}
\value{
Data frame with \code{Bin_ID} first.
}
\description{
Read the CheckM1 extended QA table
}
