% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mags.R
\name{read_checkm2}
\alias{read_checkm2}
\title{Read CheckM2 quality reports}
\usage{
read_checkm2(paths.v)
}
\arguments{
\item{paths.v}{One or more \code{quality_report}-style files (combined when several).}
}
\value{
Data frame with \code{Bin_ID} first.
}
\description{
Read CheckM2 quality reports
}
