% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc.R
\name{read_checkv_quality}
\alias{read_checkv_quality}
\title{Read CheckV quality summary}
\usage{
read_checkv_quality(path)
}
\arguments{
\item{path}{Path to \code{quality_summary.tsv}.}
}
\value{
Data frame with \code{Sequence_ID} first.
}
\description{
Read CheckV quality summary
}
