% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isolate.R
\name{read_fastani}
\alias{read_fastani}
\title{Read fastANI results}
\usage{
read_fastani(path)
}
\arguments{
\item{path}{Path to \verb{<group>.fastani.tsv} (query, reference, ANI, mapped and total fragments).}
}
\value{
Data frame with \code{Query}, \code{Reference}, \code{ANI}, \code{Mapped_fragments}, \code{Total_fragments}.
}
\description{
Read fastANI results
}
