% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc.R
\name{read_gene_catalogue_membership}
\alias{read_gene_catalogue_membership}
\title{Read the gene catalogue cluster membership table}
\usage{
read_gene_catalogue_membership(path)
}
\arguments{
\item{path}{Path to \code{gene_catalogue_membership.tsv}.}
}
\value{
Data frame.
}
\description{
Read the gene catalogue cluster membership table
}
