% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mags.R
\name{read_gtdbtk}
\alias{read_gtdbtk}
\title{Read GTDB-Tk summary tables}
\usage{
read_gtdbtk(paths.v)
}
\arguments{
\item{paths.v}{Paths to \verb{*.bac120.summary.tsv} and/or \verb{*.ar53.summary.tsv}.}
}
\value{
Data frame with \code{Bin_ID}, \code{Classification}, rank columns and \code{Classification_method}.
}
\description{
Read GTDB-Tk summary tables
}
