% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isolate.R
\name{read_isescan}
\alias{read_isescan}
\title{Read collated ISEScan results}
\usage{
read_isescan(path)
}
\arguments{
\item{path}{Path to \code{ISEScan_summary.tsv}.}
}
\value{
Data frame with \code{Genome} first.
}
\description{
Read collated ISEScan results
}
