% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isolate.R
\name{read_panaroo}
\alias{read_panaroo}
\title{Read a Panaroo gene presence/absence table}
\usage{
read_panaroo(path)
}
\arguments{
\item{path}{Path to \code{gene_presence_absence.csv}.}
}
\value{
\code{gm_profile} of presence (genes x genomes) with \code{Gene}, \code{Non_unique_gene_name} and \code{Annotation} features.
}
\description{
Read a Panaroo gene presence/absence table
}
