% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc.R
\name{read_read_stats}
\alias{read_read_stats}
\title{Read the pipeline read statistics report}
\usage{
read_read_stats(path)
}
\arguments{
\item{path}{Path to \code{read_stat_report.tsv}.}
}
\value{
Data frame with \code{Sample_ID} first.
}
\description{
Read the pipeline read statistics report
}
