% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_misc.R
\name{read_rpkm}
\alias{read_rpkm}
\title{Read a gene catalogue RPKM or read-count table}
\usage{
read_rpkm(path, value_type = "normalised_rpkm")
}
\arguments{
\item{path}{Path to a wide table with \code{Gene_ID} then one column per sample.}

\item{value_type}{Profile value type.}
}
\value{
\code{gm_profile} of genes.
}
\description{
Read a gene catalogue RPKM or read-count table
}
