% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profiling.R
\name{read_singlem_fraction}
\alias{read_singlem_fraction}
\title{Read the SingleM prokaryotic fraction table}
\usage{
read_singlem_fraction(path)
}
\arguments{
\item{path}{Path to \code{metagenome.prokaryotic_fraction.tsv}.}
}
\value{
Data frame.
}
\description{
Read the SingleM prokaryotic fraction table
}
