% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isolate.R
\name{read_tree}
\alias{read_tree}
\title{Read a phylogenetic tree}
\usage{
read_tree(path)
}
\arguments{
\item{path}{Newick file (e.g. \verb{<group>.treefile}).}
}
\value{
An \code{ape::phylo} object with tip labels stripped of file extensions.
}
\description{
Read a phylogenetic tree
}
