% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{start_project}
\alias{start_project}
\title{Start a project object}
\usage{
start_project(config.l)
}
\arguments{
\item{config.l}{A \code{gm_config} from \code{\link[=read_config]{read_config()}}.}
}
\value{
A \code{gm_project} list with \code{config}, \code{metadata}, \code{run_params}, \code{read_stats},
and empty \code{profiles} and \code{tables}.
}
\description{
Reads the metadata, run parameters and read statistics.
}
