% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{subset_features}
\alias{subset_features}
\title{Subset profile features}
\usage{
subset_features(profile, feature_ids.v)
}
\arguments{
\item{profile}{A \code{gm_profile}.}

\item{feature_ids.v}{Features to keep.}
}
\value{
\code{gm_profile}.
}
\description{
Subset profile features
}
