% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{taxonomic_profile_names}
\alias{taxonomic_profile_names}
\title{Names of profiles with taxonomic rank columns}
\usage{
taxonomic_profile_names(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project}.}
}
\value{
Character vector.
}
\description{
Names of profiles with taxonomic rank columns
}
