% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_ranks}
\alias{taxonomy_ranks}
\title{Taxonomic ranks used throughout gmaio}
\usage{
taxonomy_ranks()
}
\value{
Named character vector of GTDB rank prefixes, named by rank.
}
\description{
Taxonomic ranks used throughout gmaio
}
