% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_gmaio}
\alias{theme_gmaio}
\title{The gmaio ggplot2 theme}
\usage{
theme_gmaio(base_size = 9, legend_position = "right")
}
\arguments{
\item{base_size}{Base font size in points.}

\item{legend_position}{Legend position.}
}
\value{
A ggplot2 theme.
}
\description{
The gmaio ggplot2 theme
}
