% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{top_loadings}
\alias{top_loadings}
\title{Top ordination loadings}
\usage{
top_loadings(ordination, axis = "PC1", n = 10)
}
\arguments{
\item{ordination}{A PCA \code{gm_ordination}.}

\item{axis}{Axis name, e.g. \code{"PC1"}.}

\item{n}{Number of features from each end.}
}
\value{
Data frame of the most negative and most positive loadings.
}
\description{
Top ordination loadings
}
