% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_isolate.R
\name{write_isolate_tables}
\alias{write_isolate_tables}
\title{Write the standard isolate tables}
\usage{
write_isolate_tables(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project} in isolate mode.}
}
\value{
Paths written, invisibly.
}
\description{
Write the standard isolate tables
}
