% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metagenome.R
\name{add_palettes}
\alias{add_palettes}
\title{Assign project palettes and save them to palettes.yml}
\usage{
add_palettes(project.l, min_taxon_abundance = 0.1)
}
\arguments{
\item{project.l}{A \code{gm_project}.}

\item{min_taxon_abundance}{Only taxa reaching this relative abundance (percent) in some sample get a saved colour.}
}
\value{
Updated \code{gm_project}.
}
\description{
Colours metadata variables (group variables, discrete covariates, \code{colour_variables}
and samples) and taxa at phylum and the configured analysis ranks.
}
