% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metagenome.R
\name{add_singlem}
\alias{add_singlem}
\title{Add SingleM profiles}
\usage{
add_singlem(project.l)
}
\arguments{
\item{project.l}{A \code{gm_project}.}
}
\value{
Updated \code{gm_project}.
}
\description{
Adds \code{singlem_coverage}, \code{singlem_relative} and, when the prokaryotic fraction is
available, \code{singlem_scaled} (relative abundance x read fraction / 100, i.e. percent of
all reads).
}
