% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_profile}
\alias{get_profile}
\title{Get a profile restricted to analysis samples}
\usage{
get_profile(project.l, name, include_excluded = FALSE, renormalise = TRUE)
}
\arguments{
\item{project.l}{A \code{gm_project}.}

\item{name}{Profile name (see \code{names(project.l$profiles)}).}

\item{include_excluded}{Keep excluded samples.}

\item{renormalise}{Rescale relative abundance to 100 after removing samples.}
}
\value{
\code{gm_profile}.
}
\description{
Get a profile restricted to analysis samples
}
