% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_to_wide_df}
\alias{profile_to_wide_df}
\title{Profile as a wide data frame for export}
\usage{
profile_to_wide_df(
  profile,
  annotation_columns = c("Feature_ID", "Label"),
  sample_labels = NULL
)
}
\arguments{
\item{profile}{A \code{gm_profile}.}

\item{annotation_columns}{Feature annotation columns to include before the samples.}

\item{sample_labels}{Optional named vector mapping sample IDs to column names.}
}
\value{
Data frame.
}
\description{
Profile as a wide data frame for export
}
