% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{special_colours}
\alias{special_colours}
\title{Fixed colours for special categories}
\usage{
special_colours()
}
\value{
Named character vector for \code{"Other"}, \code{"Unassigned"}, missing values and reference genomes.
}
\description{
Fixed colours for special categories
}
